workaround for nil sender issue
Immediately after sending a message, the sender may be nil, in which case reuseIdentifier defaults to ATLIncomingMessageCellIdentifier.  This change checks for a nil sender or nil sender.userID as possible conditions for using ATLOutgoingMessageCellIdentifier.  This may not be the correct solution to fix the problem, but it seems to work for now.